fix(pandas): do not use `pandas.Series.argmin`
Prior to this commit, the following deprecation warning occurs when using
`pandas>=0.22.0`:

```
FutureWarning: 'argmin' is deprecated. Use 'idxmin' instead. The behavior of 'argmin' will be corrected to return the positional minimum in the future. Use 'series.values.argmin' to get the position of the minimum now.
```

In this commit, discontinue use of deprecated `pandas.Series.argmin` in
`ui.views.find_closest`.