fix(shapes-canvas): fix flicker on focus change
Prior to this commit, the shapes canvas view refreshes/flickers when it
gets/loses focus.  The reason for this is the way the window **expose**
event was debounced on the [_trailing_ edge][debounce].  The goal of
debouncing is to prevent triggering too many consecutive resize handling
events before the first one is allowed to complete.  However, in the case
of a change of window focus, it is ideal to apply a view refresh
immediately, otherwise there is a perceived "flicker".  Since the window
**expose** event is triggered _both_ on window **_resize_** and **_change
of focus_**, debouncing on the _trailing_ edge results in a delay (i.e.,
flicker) when a change of window focus occurs.

In this commit, call debounced `_on_expose_event` handler on _leading_
edge to make UI update more responsive when, e.g., changing window focus,
as well as on _trailing_ edge (if it occurs again).  This both prevents
too many consecutive resize events from triggering **and** refreshes the
window immediately on a change of focus event.

Fixes sci-bots/microdrop/issues/254.

[debounce]: https://lodash.com/docs#debounce